% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_stats.R
\name{summarize_stats}
\alias{summarize_stats}
\title{Generate summary statistics across groups
Outputs the mean, median, and range of a numerical variable across different groups of a categorical variable in a dataframe.}
\usage{
summarize_stats(df, numcol, categcol, NAs = TRUE)
}
\arguments{
\item{df}{dataframe containing the dataset to be analysed. Named df as a shorter form for dataframe.}

\item{numcol}{numeric vector specifying the numerical column used for summary statistics. Named numcol as a shorter form of numeric column.}

\item{categcol}{character vector specifying the categorical column with groups of interest. Named categcol as a shorter form of categorical column.}

\item{NAs}{logical specifying whether you want to exclude missing values from your summary statistic calculations. If TRUE, remove all missing values. Named NAs to represent the presence/absence of missing values.}
}
\value{
a table with the mean, median, and range values of the \code{numcol} column across all the groups of the \code{categcol} column.
}
\description{
summarize_stats generates the mean, median, and the range values of a numeric variable across a categorical variable.
This facilitates the comparison of these summary statistics across different groups within a categorical variable of interest, allowing for further downstream analyses.
}
\examples{
summarize_stats(df = datateachr::steam_games, numcol = original_price, categcol = types, NAs = TRUE)
summarize_stats(df = palmerpenguins::penguins, numcol = bill_depth_mm, categcol = sex, NAs = TRUE)
}
